[refactor] no longer to have subfolders within each category folder under UI, hence options.js moved to top level category folder
